move playlist manipulation code to module
